requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
distro==1.9.0
fastapi==0.135.3
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
joblib==1.5.3
numpy==2.4.4
openai==2.30.0
pandas==3.0.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
